+++ tests/unit/s2n_mem_allocator_test.c
@@ -125,7 +125,7 @@ void mock_client(struct s2n_test_io_pair *io_pair)
     /* Active application bytes consumed is reset to 0 in before writing data. */
     /* Its value should equal to bytes written after writing */
     ssize_t bytes_written = s2n_send(conn, buffer, i, &blocked);
-    if (bytes_written != conn->active_application_bytes_consumed) {
+    if ((uint64_t)bytes_written != conn->active_application_bytes_consumed) {
         exit(0);
     }
 
@@ -172,7 +172,7 @@ int main(int argc, char **argv)
     EXPECT_NOT_NULL(private_key_pem = malloc(S2N_MAX_TEST_PEM_SIZE));
     EXPECT_NOT_NULL(dhparams_pem = malloc(S2N_MAX_TEST_PEM_SIZE));
 
-    for (int is_dh_key_exchange = 0; is_dh_key_exchange <= 1; is_dh_key_exchange++) {
+    for (size_t is_dh_key_exchange = 0; is_dh_key_exchange <= 1; is_dh_key_exchange++) {
         struct s2n_cert_chain_and_key *chain_and_keys[SUPPORTED_CERTIFICATE_FORMATS];
 
         /* Create a pipe */
@@ -198,7 +198,7 @@ int main(int argc, char **argv)
         conn->actual_protocol_version = S2N_TLS12;
 
         EXPECT_NOT_NULL(config = s2n_config_new());
-        for (int cert = 0; cert < SUPPORTED_CERTIFICATE_FORMATS; cert++) {
+        for (size_t cert = 0; cert < SUPPORTED_CERTIFICATE_FORMATS; cert++) {
             EXPECT_SUCCESS(s2n_read_test_pem(certificate_paths[cert], cert_chain_pem, S2N_MAX_TEST_PEM_SIZE));
             EXPECT_SUCCESS(s2n_read_test_pem(private_key_paths[cert], private_key_pem, S2N_MAX_TEST_PEM_SIZE));
             EXPECT_NOT_NULL(chain_and_keys[cert] = s2n_cert_chain_and_key_new());
@@ -247,7 +247,7 @@ int main(int argc, char **argv)
         } while (shutdown_rc != 0);
 
         EXPECT_SUCCESS(s2n_connection_free(conn));
-        for (int cert = 0; cert < SUPPORTED_CERTIFICATE_FORMATS; cert++) {
+        for (size_t cert = 0; cert < SUPPORTED_CERTIFICATE_FORMATS; cert++) {
             EXPECT_SUCCESS(s2n_cert_chain_and_key_free(chain_and_keys[cert]));
         }
         EXPECT_SUCCESS(s2n_config_free(config));
